fix(DeleteLink): accounts can delete only your own links

+++ src/modules/links/useCases/deleteLink/DeleteLinkUseCase.ts
@@ -2,8 +2,11 @@ import { inject, injectable } from "tsyringe";
 
 import { ILinksRepository } from "../../repositories/ILinksRepository";
 
+import { AppError } from "../../../../shared/errors/AppErrors";
+
 interface IRequest {
     linkId: string;
+    account_id: string;
 }
 
 @injectable()
@@ -14,8 +17,12 @@ class DeleteLinkUseCase {
         private linksRepository: ILinksRepository
     ) { }
 
-    async execute({ linkId }: IRequest): Promise<void> {
+    async execute({ linkId, account_id }: IRequest): Promise<void> {
         const link = await this.linksRepository.findById(linkId);
+
+        if (link.account_id !== account_id) {
+            throw new AppError("Unauthorized", 401)
+        }
 
         await this.linksRepository.delete(link)
     }
